requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\dbots\666 Mod\bot.py: 1,2
# D:\dbots\666 Mod\cogs\admin.py: 1,2
# D:\dbots\666 Mod\cogs\hep.py: 1,2
# D:\dbots\666 Mod\cogs\mod.py: 1,2
# D:\dbots\666 Mod\cogs\secret.py: 1,2
discord.py == 1.6.0

# D:\dbots\666 Mod\bot.py: 3
# D:\dbots\666 Mod\cogs\admin.py: 6
# D:\dbots\666 Mod\cogs\hep.py: 6
# D:\dbots\666 Mod\cogs\mod.py: 9
jishaku == 1.19.1.200

# D:\dbots\666 Mod\bot.py: 9
# D:\dbots\666 Mod\fixmute.py: 1
pymongo == 3.11.0

Sphinx==3.5.3
rst2pdf==0.98
karma_sphinx_theme==0.0.8
